requirements.txt
django
djangorestframework
psycopg2-binary
requests
markdown
django-filter
drf_yasg